+++ tracker-card.js
@@ -49,6 +49,10 @@ class TrackerCard extends HTMLElement {
           table {
             width: 100%;
             padding: 0 32px 0 32px;
+            border-spacing: 0px;
+          }
+          td {
+            padding: 2px;
           }
           thead th {
             text-align: left;
@@ -81,6 +85,16 @@ class TrackerCard extends HTMLElement {
             padding-top: 10px;
             text-transform: capitalize;
           }
+          .update_pending {
+            
+          }
+          .hidden-button {
+            padding: 0px;
+            border: 0px;
+            background: none;
+            font-weight: 700;
+            color: red;
+          }
         `;
     content.innerHTML = `
       <div id='content'>
@@ -105,6 +119,7 @@ class TrackerCard extends HTMLElement {
     const config = this._config;
     const root = this.shadowRoot;
     const card = root.lastChild;
+    const pending_updates = [];
     this.myhass = hass;
     this.handlers = this.handlers || [];
     let card_content = '';
@@ -119,24 +134,52 @@ class TrackerCard extends HTMLElement {
         const list = this._filterCards(hass.states[tracker].attributes);
         const domain = hass.states[tracker].attributes.domain;
         const repo = hass.states[tracker].attributes.repo;
+        
+        for (var i in list) {
+          if (list[i][1].has_update) {
+            pending_updates.push(list[i])
+          }
+        }
+        
+        
         card_content += `
           <tr><td colspan='3' class='separator'>${domain.replace('_', ' ')}:</td></tr>
         `;
-
         if (list !== undefined && list.length > 0) {
           const updated_content = `
             ${list.map(elem => `
-
+              ${elem[1].has_update?`
+                <tr class='update_pending'>
+                  <td class='name'>
+                    <a href="${elem[1].change_log?elem[1].change_log:'#'}" target='_blank'>
+                      ${elem[0]}
+                    </a>
+                  </td>
+                  <td class='local'>
+                    ${elem[1].local?elem[1].local:'n/a'}
+                  </td>
+                  <td class='remote'>
+                    <div>
+                      <button title="Update this" class='hidden-button' id='${elem[0]}'>
+                        ${elem[1].remote?elem[1].remote:'n/a'}
+                      </button>
+                    </div>
+                  </td>
+            `:`
                 <tr>
-                  <td class='name'><a href="${elem[1].repo?elem[1].repo:'#'}" target='_blank'>${elem[0]}</a></td>
-                  <td>${elem[1].local?elem[1].local:'n/a'}</td>
-                  <td>
-                    ${elem[1].has_update?`
-                    <a href="${elem[1].change_log?elem[1].change_log:'#'}" target='_blank'>${elem[1].remote?elem[1].remote:'n/a'}</a>
-                    `:(elem[1].remote?elem[1].remote:'n/a')}
-                    </td>
-                </tr>
-            `).join('')}
+                  <td class='name'>
+                    <a href="${elem[1].repo?elem[1].repo:'#'}" target='_blank'>
+                      ${elem[0]}
+                    </a>
+                  </td>
+                  <td class='local'>
+                    ${elem[1].local?elem[1].local:'n/a'}
+                  </td>
+                  <td class='remote'>
+                    ${elem[1].remote?elem[1].remote:'n/a'}
+                  </td>
+            `}`
+            ).join('')}
           `;
           card_content += updated_content;
         }
@@ -155,6 +198,12 @@ class TrackerCard extends HTMLElement {
     });
     card_content += `</tbody></table>`;
     root.getElementById('content').innerHTML = card_content;
+    for (var i in pending_updates) {
+      card.querySelector('#' + pending_updates[i][0]).addEventListener('click', event => {
+        this.myhass.callService('custom_updater', 'install', {'element': pending_updates[i][0]});
+        this.myhass.callService('custom_updater', 'check_all', {});
+      });
+    }
 
   }
   getCardSize() {